Final upload for paper
Aggregate_Years has a few cosmetic changes, still works the same.

Distribute_Urban...etc is a function for rolling urbanization (LOCALE) data to previous years.  It's slow, I'll (try to) upload a .csv copy of the output in a bit.

Writeup for Paper is for the paper.  It needs an aggregated df (df) to work.  Probably better as an introduction.

Region_Map.png should go with 1st graph in Writeup for Paper - Thomas.